requirements.txt
Flask==2.1.2
pandas==1.4.2
numpy==1.21.5
scikit-learn==0.24.2
tensorflow==2.8.0
google-generativeai==0.1.0
gunicorn==20.1.0
